+++ app/Http/Controllers/AdminDonorsController.php
@@ -17,7 +17,7 @@ class AdminDonorsController extends Controller
         $user = Auth::user();
 
         if ($user->role->id !== 1) {
-            return redirect()->route('401')->with('error', 'No tienes acceso a esta página.');
+            return redirect()->route('401');
         }
 
         // Filtra los usuarios que tengan el rol 'coordi' y carga sus programas
@@ -32,6 +32,12 @@ class AdminDonorsController extends Controller
 
     public function store(Request $request)
     {
+        $user = Auth::user();
+
+        if ($user->role->id !== 1) {
+            return redirect()->route('401');
+        }
+
         $request->validate([
             'name' => 'required|max:60',
             'email' => 'required|email|max:60',
@@ -51,6 +57,12 @@ class AdminDonorsController extends Controller
 
     public function update(Request $request, User $admin_donor)
     {
+        $user = Auth::user();
+
+        if ($user->role->id !== 1) {
+            return redirect()->route('401');
+        }
+
         $request->validate([
             'name' => 'required|max:60',
             'email' => 'required|email|max:60',
@@ -76,6 +88,12 @@ class AdminDonorsController extends Controller
 
     public function destroy(User $admin_donor)
     {
+        $user = Auth::user();
+
+        if ($user->role->id !== 1) {
+            return redirect()->route('401');
+        }
+        
         $admin_donor->delete();
         return redirect('admin-donors');
     }
